Use an empty doctor for tests
