+++ features/bootstrap/Test/Doctor/NullDoctor.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Test\Doctor;
+
+use Dock\Doctor\Doctor;
+use Symfony\Component\Console\Output\OutputInterface;
+
+class NullDoctor implements Doctor
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function examine(OutputInterface $output, $dryRun)
+    {
+    }
+}
