+++ src/Doctor/Doctor.php
@@ -4,29 +4,11 @@ namespace Dock\Doctor;
 
 use Symfony\Component\Console\Output\OutputInterface;
 
-class Doctor
+interface Doctor
 {
     /**
-     * @var Task[]
+     * @param OutputInterface $output
+     * @param bool $dryRun
      */
-    private $tasks;
-
-    /**
-     * @param Task[] $tasks
-     */
-    public function __construct(array $tasks)
-    {
-        $this->tasks = $tasks;
-    }
-
-    public function examine(OutputInterface $output, $dryRun)
-    {
-        $output->writeLn('<info>Running Docker Doctor</info>');
-
-        foreach ($this->tasks as $task) {
-            $task->run($output, $dryRun);
-        }
-
-        $output->writeLn('<info>Yay! Your setup is working perfectly!</info>');
-    }
+    public function examine(OutputInterface $output, $dryRun);
 }
